# artifacts/api-server/src/routes/productCodes.ts
import { Router, type IRouter } from "express";
import { eq } from "drizzle-orm";
import { db, productCodesTable } from "@workspace/db";
import { CreateProductCodeBody, DeleteProductCodeParams } from "@workspace/api-zod";

const router: IRouter = Router();

router.get("/product-codes", async (req, res): Promise<void> => {
  const codes = await db
    .select()
    .from(productCodesTable)
    .orderBy(productCodesTable.productName);
  res.json(codes.map((c) => ({
    id: c.id,
    code: c.code,
    productName: c.productName,
    createdAt: c.createdAt.toISOString(),
  })));
});

router.post("/product-codes", async (req, res): Promise<void> => {
  const parsed = CreateProductCodeBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  const [created] = await db
    .insert(productCodesTable)
    .values({
      code: parsed.data.code,
      productName: parsed.data.productName,
    })
    .returning();

  res.status(201).json({
    id: created.id,
    code: created.code,
    productName: created.productName,
    createdAt: created.createdAt.toISOString(),
  });
});

router.delete("/product-codes/:id", async (req, res): Promise<void> => {
  const params = DeleteProductCodeParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }

  const [deleted] = await db
    .delete(productCodesTable)
    .where(eq(productCodesTable.id, params.data.id))
    .returning();

  if (!deleted) {
    res.status(404).json({ error: "Código não encontrado" });
    return;
  }

  res.json({ success: true });
});

export default router;