# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Atacadão system
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface AuthUser {
  username: string;
}

export interface ProductCode {
  id: number;
  code: string;
  productName: string;
  createdAt: string;
}

export interface ProductCodeInput {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  productName: string;
}

export interface StockItem {
  id: number;
  productName: string;
  /** @nullable */
  description?: string | null;
  location: string;
  createdAt: string;
  updatedAt?: string;
}

export interface StockItemInput {
  /** @minLength 1 */
  productName: string;
  description?: string;
  /** @minLength 1 */
  location: string;
}

export interface StockItemUpdate {
  /** @minLength 1 */
  productName?: string;
  description?: string;
  /** @minLength 1 */
  location?: string;
}

export type StockSummaryLocationBreakdownItem = {
  location: string;
  count: number;
};

export interface StockSummary {
  totalItems: number;
  totalCodes: number;
  locationBreakdown: StockSummaryLocationBreakdownItem[];
}