# lib/api-zod/src/generated/types/productCode.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Atacadão system
 * OpenAPI spec version: 0.1.0
 */

export interface ProductCode {
  id: number;
  code: string;
  productName: string;
  createdAt: string;
}