# lib/api-zod/src/generated/types/stockItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Atacadão system
 * OpenAPI spec version: 0.1.0
 */

export interface StockItem {
  id: number;
  productName: string;
  /** @nullable */
  description?: string | null;
  location: string;
  createdAt: string;
  updatedAt?: string;
}