# lib/api-zod/src/generated/types/stockItemUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Atacadão system
 * OpenAPI spec version: 0.1.0
 */

export interface StockItemUpdate {
  /** @minLength 1 */
  productName?: string;
  description?: string;
  /** @minLength 1 */
  location?: string;
}