# lib/api-zod/src/generated/types/stockSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Atacadão system
 * OpenAPI spec version: 0.1.0
 */
import type { StockSummaryLocationBreakdownItem } from './stockSummaryLocationBreakdownItem';

export interface StockSummary {
  totalItems: number;
  totalCodes: number;
  locationBreakdown: StockSummaryLocationBreakdownItem[];
}